# src/app/api/validation/missing-fields/route.ts
/**
 * VALIDATION API - Ministry Compliance Missing Fields Detection
 * Task #90.2 - Implement API endpoint using ministry-validation service
 * Task #101.1 - Enhanced with section breakdown and detailed missing fields
 *
 * GET /api/validation/missing-fields?developerId={uuid}&includeSections={boolean}
 *
 * Analyzes developer's properties for missing required/recommended ministerial fields
 * and format errors. Returns comprehensive validation summary with actionable warnings.
 *
 * Query Parameters:
 * - developerId: UUID of developer (optional, defaults to authenticated user)
 * - includeSections: boolean (optional, adds detailed section breakdown)
 */

import { NextRequest, NextResponse } from 'next/server'
import { createClient } from '@/lib/supabase/server'
import { validateProperties } from '@/lib/ministry-validation'
import type { ParsedProperty } from '@/lib/api-schemas'
import type {
  ValidationMissingFieldsResponse,
  MissingFieldSummaryEntry,
  PropertyValidationItem,
} from '@/lib/api-schemas'
import type { MissingFieldInfo, SectionBreakdown } from '@/lib/ministry-validation'
import { REQUIRED_FIELDS, RECOMMENDED_FIELDS } from '@/lib/ministry-validation'

export const dynamic = 'force-dynamic'

export async function GET(request: NextRequest) {
  try {
    const supabase = await createClient()

    // ========================================================================
    // AUTHENTICATION
    // ========================================================================

    const { data: { user }, error: authError } = await supabase.auth.getUser()

    if (authError || !user) {
      return NextResponse.json(
        { success: false, error: 'Unauthorized' },
        { status: 401 }
      )
    }

    // ========================================================================
    // DEVELOPER IDENTIFICATION
    // ========================================================================

    // Get query parameters
    const searchParams = request.nextUrl.searchParams
    const requestedDeveloperId = searchParams.get('developerId')
    const includeSections = searchParams.get('includeSections') === 'true'

    let developerId: string

    if (requestedDeveloperId) {
      // TODO: Add admin role check here when admin roles are implemented
      // For now, only allow users to query their own developer profile
      const { data: developer, error: devError } = await supabase
        .from('developers')
        .select('id, user_id')
        .eq('id', requestedDeveloperId)
        .single()

      if (devError || !developer) {
        return NextResponse.json(
          { success: false, error: 'Developer not found' },
          { status: 404 }
        )
      }

      // Security check: Ensure user owns this developer profile
      if (developer.user_id !== user.id) {
        return NextResponse.json(
          { success: false, error: 'Forbidden - not your developer profile' },
          { status: 403 }
        )
      }

      developerId = developer.id
    } else {
      // Use authenticated user's developer profile
      const { data: developer, error: devError } = await supabase
        .from('developers')
        .select('id')
        .eq('user_id', user.id)
        .single()

      if (devError || !developer) {
        return NextResponse.json(
          { success: false, error: 'Developer profile not found' },
          { status: 404 }
        )
      }

      developerId = developer.id
    }

    // ========================================================================
    // FETCH PROPERTIES WITH RAW CSV DATA
    // ========================================================================

    const { data: dbProperties, error: propsError } = await supabase
      .from('properties')
      .select(`
        id,
        developer_id,
        apartment_number,
        property_type,
        wojewodztwo,
        powiat,
        gmina,
        miejscowosc,
        ulica,
        nr_budynku,
        kod_pocztowy,
        price_per_m2,
        base_price,
        final_price,
        area,
        rooms,
        floor,
        status,
        raw_csv_data(raw_data)
      `)
      .eq('developer_id', developerId)
      .order('created_at', { ascending: false })

    if (propsError) {
      console.error('❌ VALIDATION API: Error fetching properties:', propsError)
      return NextResponse.json(
        { success: false, error: propsError.message },
        { status: 500 }
      )
    }

    // Handle empty properties list
    if (!dbProperties || dbProperties.length === 0) {
      return NextResponse.json({
        success: true,
        data: {
          summary: {
            totalProperties: 0,
            propertiesWithIssues: 0,
            propertiesValid: 0,
            complianceScore: 100, // No properties = 100% compliant
          },
          missingFieldsSummary: {},
          properties: [],
        },
      } as ValidationMissingFieldsResponse)
    }

    // ========================================================================
    // CONVERT DATABASE PROPERTIES TO PARSED PROPERTY FORMAT
    // ========================================================================

    const parsedProperties: ParsedProperty[] = dbProperties.map((dbProp) => {
      // Extract raw CSV data (from raw_csv_data relation)
      const rawCsvData = Array.isArray(dbProp.raw_csv_data) && dbProp.raw_csv_data.length > 0
        ? (dbProp.raw_csv_data[0] as { raw_data?: Record<string, unknown> })?.raw_data || {}
        : {}

      return {
        // Basic identification
        property_number: dbProp.apartment_number || undefined,
        property_type: dbProp.property_type || undefined,

        // Location (from database columns)
        wojewodztwo: dbProp.wojewodztwo || undefined,
        powiat: dbProp.powiat || undefined,
        gmina: dbProp.gmina || undefined,
        miejscowosc: dbProp.miejscowosc || undefined,
        ulica: dbProp.ulica || undefined,
        numer_nieruchomosci: dbProp.nr_budynku || undefined,
        kod_pocztowy: dbProp.kod_pocztowy || undefined,

        // Pricing (from database columns)
        price_per_m2: dbProp.price_per_m2 || undefined,
        base_price: dbProp.base_price || undefined,
        total_price: dbProp.final_price || undefined,
        area: dbProp.area || undefined,

        // Additional fields
        liczba_pokoi: dbProp.rooms || undefined,
        kondygnacja: dbProp.floor || undefined,

        // Raw CSV data (contains ALL ministry fields including developer info)
        raw_data: rawCsvData as Record<string, unknown>,
      }
    })

    // ========================================================================
    // VALIDATE PROPERTIES USING MINISTRY VALIDATION SERVICE
    // ========================================================================

    const validationResult = validateProperties(parsedProperties)

    // ========================================================================
    // BUILD MISSING FIELDS SUMMARY
    // ========================================================================

    const missingFieldsSummary: Record<string, MissingFieldSummaryEntry> = {}

    // Count how many properties are missing each field
    const fieldCounts: Record<string, { count: number; severity: 'critical' | 'warning' | 'info' }> = {}

    validationResult.propertyResults.forEach((propResult) => {
      // Count required fields
      propResult.missingRequired.forEach((field) => {
        if (!fieldCounts[field]) {
          fieldCounts[field] = { count: 0, severity: 'critical' }
        }
        fieldCounts[field].count++
      })

      // Count recommended fields
      propResult.missingRecommended.forEach((field) => {
        if (!fieldCounts[field]) {
          fieldCounts[field] = { count: 0, severity: 'info' }
        }
        fieldCounts[field].count++
      })
    })

    // Build summary entries with percentages and labels
    Object.entries(fieldCounts).forEach(([fieldName, { count, severity }]) => {
      const percentage = Math.round((count / dbProperties.length) * 100)

      // Get human-readable field label
      let fieldLabel = fieldName
      if (REQUIRED_FIELDS[fieldName as keyof typeof REQUIRED_FIELDS]) {
        fieldLabel = REQUIRED_FIELDS[fieldName as keyof typeof REQUIRED_FIELDS]
      } else if (RECOMMENDED_FIELDS[fieldName as keyof typeof RECOMMENDED_FIELDS]) {
        fieldLabel = RECOMMENDED_FIELDS[fieldName as keyof typeof RECOMMENDED_FIELDS]
      }

      missingFieldsSummary[fieldName] = {
        count,
        percentage,
        severity,
        fieldLabel,
      }
    })

    // ========================================================================
    // BUILD PROPERTIES RESPONSE
    // ========================================================================

    const properties: PropertyValidationItem[] = validationResult.propertyResults.map((propResult, index) => {
      const dbProp = dbProperties[index]

      // Format address for display
      const addressParts = [
        dbProp.miejscowosc,
        dbProp.ulica,
        dbProp.nr_budynku,
      ].filter(Boolean)

      const address = addressParts.length > 0
        ? addressParts.join(', ')
        : `Nieruchomość ${propResult.propertyNumber}`

      return {
        id: dbProp.id,
        propertyNumber: propResult.propertyNumber,
        address,
        status: propResult.valid ? 'valid' : 'invalid',
        errors: propResult.errors,
        warnings: propResult.warnings,
        missingRequired: propResult.missingRequired,
        missingRecommended: propResult.missingRecommended,
        invalidFormats: propResult.invalidFormats,
      }
    })

    // ========================================================================
    // BUILD FINAL RESPONSE (Task #101.1 - Enhanced)
    // ========================================================================

    // Base response structure
    const responseData: {
      summary: {
        totalProperties: number
        propertiesWithIssues: number
        propertiesValid: number
        complianceScore: number
      }
      missingFieldsSummary: Record<string, MissingFieldSummaryEntry>
      properties: PropertyValidationItem[]
      sectionBreakdown?: {
        developer: SectionBreakdown
        location: SectionBreakdown
        pricing: SectionBreakdown
        technical: SectionBreakdown
      }
      detailedMissingFields?: Array<{
        fieldName: string
        displayName: string
        category: 'required' | 'recommended' | 'developer'
        section: 'developer' | 'location' | 'pricing' | 'technical'
        severity: 'critical' | 'warning'
      }>
    } = {
      summary: {
        totalProperties: validationResult.totalProperties,
        propertiesWithIssues: validationResult.invalidProperties,
        propertiesValid: validationResult.validProperties,
        complianceScore: validationResult.complianceScore,
      },
      missingFieldsSummary,
      properties,
    }

    // Add section breakdown if requested (Task #101.1)
    if (includeSections && validationResult.sectionBreakdown) {
      responseData.sectionBreakdown = validationResult.sectionBreakdown

      // Aggregate detailed missing fields from all properties
      const allMissingFields = new Map<string, MissingFieldInfo>()

      validationResult.propertyResults.forEach((propResult) => {
        if (propResult.missingFieldsDetailed) {
          propResult.missingFieldsDetailed.forEach((field) => {
            if (!allMissingFields.has(field.fieldName)) {
              allMissingFields.set(field.fieldName, field)
            }
          })
        }
      })

      responseData.detailedMissingFields = Array.from(allMissingFields.values())
    }

    const response: ValidationMissingFieldsResponse = {
      success: true,
      data: responseData,
    }

    console.log(`✅ VALIDATION API: Analyzed ${validationResult.totalProperties} properties, compliance: ${validationResult.complianceScore}%`)

    return NextResponse.json(response)

  } catch (error: unknown) {
    console.error('❌ VALIDATION API: Unexpected error:', error)
    return NextResponse.json(
      {
        success: false,
        error: error instanceof Error ? error.message : 'Internal server error',
      },
      { status: 500 }
    )
  }
}